requirements.txt
scipy==1.4.1
numpy==1.18.2
numba==0.49.0
progressbar33==2.4
